requirements.txt
torch
torchvision
numpy
matplotlib
pyyaml